+++ TDEngine2/include/editor/Inspectors.h
@@ -67,6 +67,7 @@ namespace TDEngine2
 			TDE2_API static void DrawMeshAnimatorGUI(const TEditorContext& editorContext);
 			TDE2_API static void Draw9SliceImageGUI(const TEditorContext& editorContext);
 			TDE2_API static void DrawGridGroupLayoutGUI(const TEditorContext& editorContext);
+			TDE2_API static void DrawToggleGUI(const TEditorContext& editorContext);
 			TDE2_API static void DrawPerspectiveCameraGUI(const TEditorContext& editorContext);
 			TDE2_API static void DrawOrthographicCameraGUI(const TEditorContext& editorContext);