+++ TDEngine2/source/ecs/CComponentManager.cpp
@@ -18,6 +18,7 @@
 #include "../../include/graphics/UI/CLabelComponent.h"
 #include "../../include/graphics/UI/C9SliceImageComponent.h"
 #include "../../include/graphics/UI/GroupLayoutComponents.h"
+#include "../../include/graphics/UI/CToggleComponent.h"
 #include "../../include/graphics/animation/CAnimationContainerComponent.h"
 #include "../../include/graphics/animation/CMeshAnimatorComponent.h"
 #include "../../include/physics/2D/CBoxCollisionObject2D.h"
@@ -460,6 +461,7 @@ namespace TDEngine2
 			CreateLabelFactory,
 			Create9SliceImageFactory,
 			CreateGridGroupLayoutFactory,
+			CreateToggleFactory,
 			CreateLODStrategyComponentFactory,
 			CreateObjIdComponentFactory,
 #if TDE2_EDITORS_ENABLED